Ensure data sources are enrolled before exiting manage data sources script (#26)
The Immuta API v1 endpoint(s) for enrolling data sources is async i.e. not all data sources may be in the Immuta database by the time the Immuta API endpoint(s) return to the caller.

This PR modifies the script for enrolling data sources so that it does not exit until all data sources are fully enrolled i.e. is in the Immuta database. This is particularly important when running other processes that depend on a data source being in the Immuta database to function properly e.g. the tagging script.

Testing Strategy: Manual test - enroll a new dataset and check that all data sources are enrolled when the script exits.

Co-authored-by: Timothy Fei Truong Jr <timothy.truong@flatiron.com>